feat(api): add speech-to-text strategy
